rename bggraph.ts to graph.ts
