+++ src/__tests__/behavior-graph.test.ts
